WEBUI-1987: Move inactivity tests to a focused behavior test file
Follow-up to extracting NuxeoInactivityBehavior: move the inactivity-timer and
reactive-401-logout suites out of test/nuxeo-app.test.js into a focused
test/nuxeo-inactivity-behavior.test.js. The behavior is now exercised through a
minimal test-only host (a keepAlive <nuxeo-resource> + _logout() + lifecycle
wiring) instead of the full nuxeo-app fixture, as suggested in review.

Same assertions, same coverage; nuxeo-app.test.js keeps only the _logout() test
that belongs to nuxeo-app. All unit tests pass (2254).

(cherry picked from commit 6e273864bc62ebb5ca9920526a3929d353cf7a07)
