+++ test/nuxeo-app.test.js
@@ -402,210 +402,6 @@ suite('nuxeo-app', () => {
     app.$.nxcon.url = prev;
   });
 
-  suite('inactivity timer (WEBUI-1987)', () => {
-    const ACTIVITY_KEY = 'nuxeo-ui-inactivity-last-activity';
-    let getStub;
-    let timeoutStub;
-    let clearStub;
-    let scheduled;
-
-    const lastScheduled = () => scheduled[scheduled.length - 1];
-
-    setup(() => {
-      // drop the timer/listeners armed by the fixture's real ready() so this suite starts clean
-      app._teardownInactivityTimer();
-      app._loggingOut = false;
-      window.localStorage.removeItem(ACTIVITY_KEY); // start with no shared cross-tab activity
-      getStub = sinon.stub(config, 'get');
-      scheduled = [];
-      // Stub the timer primitives instead of faking global time (faking time can freeze the shared
-      // test page for every other suite). Capture scheduled callbacks and invoke them manually.
-      timeoutStub = sinon.stub(window, 'setTimeout').callsFake((fn, delay) => {
-        scheduled.push({ fn, delay });
-        return scheduled.length;
-      });
-      clearStub = sinon.stub(window, 'clearTimeout');
-    });
-
-    teardown(() => {
-      timeoutStub.restore();
-      clearStub.restore();
-      getStub.restore();
-      app._teardownInactivityTimer();
-      window.localStorage.removeItem(ACTIVITY_KEY);
-    });
-
-    test('arms a timeout for the configured idle period and redirects to logout when it fires', () => {
-      getStub.withArgs('session.timeout', 60).returns(1); // 1 minute
-      const redirect = sinon.stub(app, '_redirect');
-      const prev = app.$.nxcon.url;
-      app.$.nxcon.url = 'https://server/nuxeo';
-      app._setupInactivityTimer();
-      expect(scheduled).to.have.lengthOf(1);
-      expect(scheduled[0].delay).to.equal(60000);
-      window.localStorage.removeItem(ACTIVITY_KEY); // no tab has been active → real logout
-      scheduled[0].fn(); // simulate the idle period elapsing
-      expect(redirect).to.have.been.calledOnceWith('https://server/nuxeo/logout');
-      app.$.nxcon.url = prev;
-      redirect.restore();
-    });
-
-    test('user activity re-arms the timer', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      app._setupInactivityTimer();
-      expect(scheduled).to.have.lengthOf(1);
-      app._lastInactivityReset = Date.now() - 2000; // bypass the 1s throttle
-      window.dispatchEvent(new MouseEvent('mousedown', { bubbles: true }));
-      expect(clearStub).to.have.been.called;
-      expect(scheduled).to.have.lengthOf(2);
-    });
-
-    test('local activity records a shared timestamp for other tabs', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      app._setupInactivityTimer(); // initial arm propagates activity
-      expect(Number(window.localStorage.getItem(ACTIVITY_KEY))).to.be.greaterThan(0);
-    });
-
-    test('activity in another tab (storage event) re-arms this tab', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      app._setupInactivityTimer();
-      expect(scheduled).to.have.lengthOf(1);
-      // Simulate another tab writing activity; the throttle must not swallow a remote signal.
-      app._lastInactivityReset = Date.now() - 2000;
-      window.dispatchEvent(new StorageEvent('storage', { key: ACTIVITY_KEY, newValue: String(Date.now()) }));
-      expect(scheduled).to.have.lengthOf(2); // re-armed from the remote activity
-    });
-
-    test('does NOT log out when another tab was active within the timeout', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      const redirect = sinon.stub(app, '_redirect');
-      app._setupInactivityTimer();
-      window.localStorage.setItem(ACTIVITY_KEY, String(Date.now())); // another tab active just now
-      lastScheduled().fn(); // this idle tab's timer fires
-      expect(redirect).not.to.have.been.called; // re-armed instead of logging out
-      redirect.restore();
-    });
-
-    test('logs out only when all tabs have been idle for the full period', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      const redirect = sinon.stub(app, '_redirect');
-      const prev = app.$.nxcon.url;
-      app.$.nxcon.url = 'https://server/nuxeo';
-      app._setupInactivityTimer();
-      window.localStorage.setItem(ACTIVITY_KEY, String(Date.now() - 120000)); // last activity 2 min ago
-      lastScheduled().fn();
-      expect(redirect).to.have.been.calledOnceWith('https://server/nuxeo/logout');
-      app.$.nxcon.url = prev;
-      redirect.restore();
-    });
-
-    test('logs out on resume when idle beyond the timeout (e.g. after the machine slept)', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      const redirect = sinon.stub(app, '_redirect');
-      const prev = app.$.nxcon.url;
-      app.$.nxcon.url = 'https://server/nuxeo';
-      app._setupInactivityTimer();
-      // Simulate elapsed real time while the tab was hidden (both local and shared references are stale).
-      app._lastActivityTs = Date.now() - 120000;
-      window.localStorage.setItem(ACTIVITY_KEY, String(Date.now() - 120000));
-      app._checkInactivityOnResume();
-      expect(redirect).to.have.been.calledOnceWith('https://server/nuxeo/logout');
-      app.$.nxcon.url = prev;
-      redirect.restore();
-    });
-
-    test('does not log out on resume when activity was recent; re-arms instead', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      const redirect = sinon.stub(app, '_redirect');
-      app._setupInactivityTimer();
-      app._lastActivityTs = Date.now();
-      window.localStorage.setItem(ACTIVITY_KEY, String(Date.now()));
-      const before = scheduled.length;
-      app._checkInactivityOnResume();
-      expect(redirect).not.to.have.been.called;
-      expect(scheduled.length).to.be.greaterThan(before); // re-armed for the remaining time
-      redirect.restore();
-    });
-
-    test('renews the server session on activity, throttled to the keep-alive window', () => {
-      getStub.withArgs('session.timeout', 60).returns(1); // 1 min -> keep-alive window 30s
-      const execute = app.$.keepAlive.execute; // stubbed in the outer setup
-      execute.resetHistory();
-      app._setupInactivityTimer();
-      expect(execute).to.have.been.called; // initial arm renews the session
-      const afterSetup = execute.callCount;
-
-      // Activity within the keep-alive window must NOT ping again.
-      app._lastInactivityReset = Date.now() - 2000; // bypass the 1s activity throttle
-      window.dispatchEvent(new MouseEvent('mousedown'));
-      expect(execute.callCount).to.equal(afterSetup);
-
-      // Once the keep-alive window has elapsed, activity pings again.
-      app._lastKeepAlive = Date.now() - 40000; // older than the 30s window
-      app._lastInactivityReset = Date.now() - 2000;
-      window.dispatchEvent(new MouseEvent('mousedown'));
-      expect(execute.callCount).to.equal(afterSetup + 1);
-    });
-
-    test('re-running setup is idempotent (tears down before re-arming)', () => {
-      getStub.withArgs('session.timeout', 60).returns(1);
-      app._setupInactivityTimer();
-      clearStub.resetHistory();
-      app._setupInactivityTimer();
-      expect(clearStub).to.have.been.called; // teardown ran before the second arm
-    });
-
-    test('a non-positive timeout disables the feature', () => {
-      getStub.withArgs('session.timeout', 60).returns(0);
-      app._setupInactivityTimer();
-      expect(app._inactivityTimeoutMs).to.equal(0);
-      expect(scheduled).to.have.lengthOf(0);
-    });
-
-    test('converts the configured minutes into milliseconds', () => {
-      getStub.withArgs('session.timeout', 60).returns(60);
-      app._setupInactivityTimer();
-      expect(app._inactivityTimeoutMs).to.equal(60 * 60000);
-      expect(scheduled[0].delay).to.equal(60 * 60000);
-    });
-  });
-
-  suite('reactive logout on unauthorized request (WEBUI-1987)', () => {
-    let redirect;
-    let prevUrl;
-
-    setup(() => {
-      redirect = sinon.stub(app, '_redirect');
-      prevUrl = app.$.nxcon.url;
-      app.$.nxcon.url = 'https://server/nuxeo';
-      app._loggingOut = false;
-    });
-
-    teardown(() => {
-      app.$.nxcon.url = prevUrl;
-      redirect.restore();
-    });
-
-    test('redirects to logout when a 401 unauthorized-request is dispatched', () => {
-      // The fixture's attached() already wired the listener via _setupUnauthorizedRedirect().
-      document.dispatchEvent(new CustomEvent('unauthorized-request', { bubbles: true, composed: true }));
-      expect(redirect).to.have.been.calledOnceWith('https://server/nuxeo/logout');
-    });
-
-    test('is one-shot: a burst of 401s triggers a single logout redirect', () => {
-      document.dispatchEvent(new CustomEvent('unauthorized-request'));
-      document.dispatchEvent(new CustomEvent('unauthorized-request'));
-      document.dispatchEvent(new CustomEvent('unauthorized-request'));
-      expect(redirect).to.have.been.calledOnce;
-    });
-
-    test('teardown removes the 401 listener', () => {
-      app._teardownUnauthorizedRedirect();
-      document.dispatchEvent(new CustomEvent('unauthorized-request'));
-      expect(redirect).not.to.have.been.called;
-    });
-  });
-
   test('_moveDocumentsToContainer configures operation and toasts on success', async () => {
     sinon.stub(app.$.moveDocumentsOp, 'execute').resolves();
     sinon.stub(app, '_toast');
